Version bump for NPM publication.
Need to get this into production for my own servers. Next major commit will be a
rewrite of the `README.md`, which ought to take hours and hours. Checking in now
with a major new minor (not micro) number.

I've done a horrible job with versions on this project. I'm very sorry.