Convert all tests to Proof.
Converted the parser tests to Proof. Many of them were redundant. They
were created at a time when I thought that it would be reasonable to
implement all the examples in the AWS documentation.

I suppose it didn't seem like such a great idea somewhere in the middle
of the "d"s.

Keeping them, though. Tests are tests. But many have the same XML
structure, while some have identical XML.