Populate `.npmignore`.
The `.npmignore` file was blank. That was so that the `.gitignore` would
not be used, since it excluded the JavaScript compiled from
CoffeeScript. We want compiled JavaScript in our NPM distribution.

We no longer use CoffeeScript. However, we do still want to keep a
separate `.npmignore` to ignore the test directory, other directories
that are only useful in the development environment.

We neglected to add `.DS_Store` and `.AppleDouble` to `.npmignore`. This
caused problems in version 0.1.0 on NPM.

Related #9.
Closes #23.